acme: increase timer resolution to 10ms
We ran for a long time with 10ms kernel resolution,
so 10ms user space resolution here should be fine.
Some systems actually provide 1ms sleeps, which
makes this polling use a bit more cpu than we'd like.
Since the timers are for user-visible things, 10ms should
still be far from noticeable.

Reduces acme's cpu usage on Macs when plumber is missing
(and plumbproc is sleeping waiting for it to appear).

LGTM=aram, r
R=r, aram
https://codereview.appspot.com/99570043